+++ routes/web.php
@@ -4,10 +4,14 @@ use App\Http\Controllers\AboutController;
 use App\Http\Controllers\AlertController;
 use App\Http\Controllers\AlertTransportController;
 use App\Http\Controllers\Auth\SocialiteController;
+use App\Http\Controllers\DeviceLinkMapDataController;
 use App\Http\Controllers\Maps\CustomMapBackgroundController;
 use App\Http\Controllers\Maps\CustomMapController;
 use App\Http\Controllers\Maps\CustomMapDataController;
 use App\Http\Controllers\Maps\DeviceDependencyController;
+use App\Http\Controllers\Maps\GeoMapDataController;
+use App\Http\Controllers\Maps\MapDataController;
+use App\Http\Controllers\Maps\ServicesMapDataController;
 use App\Http\Controllers\PushNotificationController;
 use App\Http\Controllers\ValidateController;
 use App\Http\Middleware\AuthenticateGraph;
@@ -95,10 +99,10 @@ Route::middleware(['auth'])->group(function () {
         Route::get('custom/{map}/data', [CustomMapDataController::class, 'get'])->name('maps.custom.data');
         Route::post('custom/{map}/data', [CustomMapDataController::class, 'save'])->name('maps.custom.data.save');
         Route::get('devicedependency', 'DeviceDependencyController@dependencyMap');
-        Route::post('getdevices', 'MapDataController@getDevices')->name('maps.getdevices');
-        Route::post('getdevicelinks', 'MapDataController@getDeviceLinks')->name('maps.getdevicelinks');
-        Route::post('getgeolinks', 'MapDataController@getGeographicLinks')->name('maps.getgeolinks');
-        Route::post('getservices', 'MapDataController@getServices')->name('maps.getservices');
+        Route::post('getdevices', [MapDataController::class, 'getDevices'])->name('maps.getdevices');
+        Route::post('getdevicelinks', [DeviceLinkMapDataController::class, 'getDeviceLinks'])->name('maps.getdevicelinks');
+        Route::post('getgeolinks', [GeoMapDataController::class, 'getGeographicLinks'])->name('maps.getgeolinks');
+        Route::post('getservices', [ServicesMapDataController::class, 'getServices'])->name('maps.getservices');
     });
     Route::get('maps/devicedependency', [DeviceDependencyController::class, 'dependencyMap']);
